Move chart logic to separate template
