+++ dashboard/client/charts/charts.html
@@ -1,0 +1,30 @@
+<template name="dashboardCharts">
+  <div class="row">
+    <div class="col-md-12">
+      {{ > dashboardChartsFiltering }}
+    </div>
+  </div>
+  <div class="row">
+    <div class="col-md-12">
+      <div>
+        {{ > requestsOverTimeChart }}
+      </div>
+      <div>
+        {{ > overviewChartChart }}
+      </div>
+    </div>
+  </div>
+  <div class="row">
+    <div class="col-md-6">
+      {{ > statusCodeCountsChart }}
+    </div>
+    <div class="col-md-6">
+      {{ > responseTimeDistributionChart }}
+    </div>
+  </div>
+  <div class="row">
+    <div class="padding-sides">
+      {{ > dashboardDataTable tableDataSet=tableDataSet }}
+    </div>
+  </div>
+</template>
